chore: adding modules & hooks on blogs
