+++ modules/blogs/components/MenuListCategory.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import React, { useEffect, useState } from 'react'
 import axios from 'axios'
-import Loading from '../../components/Loading'
+import Loading from '../../../components/Loading'
 import Mapping from '@/common/utils/mapping'
 
 interface Category {
